created modules folder and change game id
